+++ src/DotNext.Tests/OptionalTests.cs
@@ -1,5 +1,7 @@
 ﻿namespace DotNext;
 
+using Runtime.CompilerServices;
+
 public sealed class OptionalTest : Test
 {
     [Fact]
@@ -270,7 +272,7 @@ public sealed class OptionalTest : Test
     [Fact]
     public static void ConvertNoneToValueType()
     {
-        var value = Optional<int>.None.Convert<string>(static i => i.ToString());
+        var value = Optional<int>.None.Convert(static i => i.ToString());
         False(value.HasValue);
         False(value.IsNull);
     }
@@ -278,7 +280,7 @@ public sealed class OptionalTest : Test
     [Fact]
     public static void ConvertValueTypeToValueType()
     {
-        var value = new Optional<int>(42).Convert<int>(static i => i + 1);
+        var value = new Optional<int>(42).Convert(static i => i + 1);
         True(value.HasValue);
         Equal(43, value.Value);
     }
@@ -286,7 +288,7 @@ public sealed class OptionalTest : Test
     [Fact]
     public static unsafe void ConvertNullToValueType()
     {
-        var value = new Optional<string>(null).Convert<int>(&int.Parse);
+        var value = new Optional<string>(null).Convert(&int.Parse);
         False(value.HasValue);
         False(value.IsNull);
     }
@@ -294,7 +296,7 @@ public sealed class OptionalTest : Test
     [Fact]
     public static void ConvertNullToRefType()
     {
-        var value = new Optional<string>(null).Convert<string>(Converter.Identity<string, string>());
+        var value = new Optional<string>(null).Convert(Converter.Identity<string, string>());
         False(value.HasValue);
         True(value.IsNull);
     }
@@ -302,8 +304,21 @@ public sealed class OptionalTest : Test
     [Fact]
     public static void ConvertNoneToRefType()
     {
-        var value = Optional<string>.None.Convert<string>(Converter.Identity<string, string>());
+        var value = Optional<string>.None.Convert(Converter.Identity<string, string>());
         False(value.HasValue);
         False(value.IsNull);
+    }
+
+    [Fact]
+    public static void OptionalToDelegate()
+    {
+        IFunctional<Func<object>> functional = Optional.None<object>();
+        Null(functional.ToDelegate().Invoke());
+
+        functional = new Optional<int>(42);
+        Equal(42, functional.ToDelegate().Invoke());
+
+        functional = Optional.None<int>();
+        Null(functional.ToDelegate().Invoke());
     }
 }