+++ src/DotNext/UserDataStorage.cs
@@ -121,13 +121,16 @@ public readonly ref partial struct UserDataStorage
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private BackingStorage? GetStorage()
-        => source is not null && (source is BackingStorage storage || GetStorage(source).TryGetValue(source, out storage!)) ? storage : null;
+        => source is not null &&
+           (source is BackingStorage storage || (Volatile.Read(in GetPartition(source))?.TryGetValue(source, out storage!) ?? false))
+            ? storage
+            : null;
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private BackingStorage GetOrCreateStorage()
     {
         if (source is not BackingStorage storage)
-            storage = GetStorage(source).GetOrCreateValue(source);
+            storage = GetOrCreatePartition(source).GetOrCreateValue(source);
 
         return storage;
     }
@@ -377,12 +380,17 @@ public readonly ref partial struct UserDataStorage
                 break;
         }
 
-        if (GetStorage() is { } source)
+        if (GetStorage() is not { } source)
         {
-            if (obj is BackingStorage destination)
-                source.CopyTo(destination);
-            else
-                GetStorage(obj).AddOrUpdate(obj, source.Copy());
+            Volatile.Read(in GetPartition(obj))?.Remove(obj);
+        }
+        else if (obj is BackingStorage destination)
+        {
+            source.CopyTo(destination);
+        }
+        else
+        {
+            GetOrCreatePartition(obj).AddOrUpdate(obj, source.Copy());
         }
     }