+++ CHANGELOG.md
@@ -1,6 +1,35 @@
 Release Notes
 ====
 
+# 08-01-2024
+<a href="https://www.nuget.org/packages/dotnext/5.11.0">DotNext 5.11.0</a>
+* Added `DotNext.Threading.Epoch` for epoch-based reclamation
+* Fixed one-shot FNV1a hashing method
+* Fixed [248](https://github.com/dotnet/dotNext/issues/248)
+* Minor performance improvements
+
+<a href="https://www.nuget.org/packages/dotnext.metaprogramming/5.11.0">DotNext.Metaprogramming 5.11.0</a>
+* Minor performance improvements
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.unsafe/5.11.0">DotNext.Unsafe 5.11.0</a>
+* Minor performance improvements
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.threading/5.11.0">DotNext.Threading 5.11.0</a>
+* Fixed `AsyncSharedLock.Downgrade` behavior, so it can be used to release a weak lock
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.io/5.11.0">DotNext.IO 5.11.0</a>
+* Minor performance improvements
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.net.cluster/5.11.0">DotNext.Net.Cluster 5.11.0</a>
+* Updated dependencies
+
+<a href="https://www.nuget.org/packages/dotnext.aspnetcore.cluster/5.11.0">DotNext.AspNetCore.Cluster 5.11.0</a>
+* Updated dependencies
+
 # 07-15-2024
 <a href="https://www.nuget.org/packages/dotnext/5.8.0">DotNext 5.8.0</a>
 * Added `FirstOrNone` and `LastOrNone` extension methods back from .NEXT 4.x as requested in [247](https://github.com/dotnet/dotNext/issues/247)